+++ CHANGELOG.md
@@ -28,6 +28,7 @@ _This release is scheduled to be released on 2024-04-01._
 - [chore] Update dependencies including electron to v29 (#3357) and node-ical
 - Updated translations for estonian (#3371)
 - Update electron to v29 and update other dependencies
+- [calendar] fullDay events over several days now show the left days from the first day on and 'today' on the last day
 - Updated layout of current weather indoor values
 
 ### Fixed